feat: ensure cart product quantity can be decreased

+++ src/actions/decrease-cart-product-quantity/index.ts
@@ -12,12 +12,12 @@ import { cartItemTable } from "@/db/schema";
 
 import { auth } from "@/lib/auth";
 
-import { removeProductFromCartSchema } from "./schema";
+import { decreaseCartProductQuantitySchema } from "./schema";
 
-export const removeProductFromCart = async (
-  data: z.infer<typeof removeProductFromCartSchema>,
+export const decreaseCartProductQuantity = async (
+  data: z.infer<typeof decreaseCartProductQuantitySchema>,
 ) => {
-  removeProductFromCartSchema.parse(data);
+  decreaseCartProductQuantitySchema.parse(data);
 
   const session = await auth.api.getSession({
     headers: await headers(),
@@ -45,5 +45,17 @@ export const removeProductFromCart = async (
     throw new Error("Unauthorized");
   }
 
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+
+    return;
+  }
+
+  await db
+
+    .update(cartItemTable)
+
+    .set({ quantity: cartItem.quantity - 1 })
+
+    .where(eq(cartItemTable.id, cartItem.id));
 };
